+++ src/components/modal/ConfirmModal/PostDeleteModal.jsx
@@ -1,4 +1,4 @@
-import { ConfirmModal, Btnwrap } from './ConfirmModalStyle';
+import { ConfirmModal, Btnwrap } from './confirmModalStyle';
 
 const PostDeleteModal = () => {
   return (
