+++ src/pages/Login/SignUp/signUpStyle.jsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+const SignUpContDiv = styled.div`
+  background-color: white;
+  min-height: 614px;
+  padding: 0 20px;
+  margin-top: 20px;
+
+  h2 {
+    font-size: 24px;
+    text-align: center;
+    margin-bottom: 40px;
+  }
+`;
+
+export default SignUpContDiv;
